Move toast message back to container element/refac
