+++ app/javascript/react/components/App.js
@@ -1,8 +1,8 @@
 import React from 'react'
+import { BrowserRouter, Route, Routes } from "react-router-dom"
 import Homepage from './Homepage.js'
 import ProfileContainer from './ProfileContainer.js'
 import SearchContainer from './SearchContainer.js'
-import { BrowserRouter, Route, Routes } from "react-router-dom"
 
 export const App = props => {
   return (
